requirements.txt
joblib==1.1.0
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
scikit_learn==1.0.2
seaborn==0.11.2
streamlit==1.6.0
tensorflow-cpu==2.8.0
